refactor(entity): split ListEntity per-entity components into list-entity/
ListEntity.tsx had grown to ~1140 lines mixing per-entity body/wide
content, the three layout shells, shared types/hooks, and the top-level
component. Split into list-entity/ with one file per entity (email,
channel, task, call, automation), one file per layout (narrow,
narrow-inbox, wide), and a shared module for LayoutProps, the layout
context/provider, useCharacterCount, hasSearchContentHits,
firstContentHit, and InboxDivider.

ListEntity.tsx now contains only ListEntity itself, MaybeEntityRow, the
ListEntityProps type, and a re-export of ListLayoutProvider — no
behavior change.
